feat: further reduce Action size by double indirection in slices
Instead of storing Rust's "fat-slice" (pointer + size) in Action,
we store just a pointer, which points to "fat-slice". This way the
size of the whole Action type is reduced from 3 x ptr_size to 2x,
at the slight performance cost of additional pointer dereference
when the value of an action is being read.